Added OnVehicleDeath and OnVehicleSpawn support
With the ability to shoot unoccupied vehicles, under certain circumstances the vehicle wont re-spawn (OnVehicleDeath and OnVehicleSpawn will never get called)

To reproduce this issue simply shoot and destroy a vehicle that has never had a player inside it since it was created/spawned/re-spawned and the vehicle will explode and will not re-spawn.  

This update adds support and vehicles will now re-spawn and will call the appropriate callbacks.