Improved Water Splash Damage
There was a velocity bug caused by animations `PED    FALL_SKYDIVE` and 'PARACHUTE    FALL_SKYDIVE` , the player's velocity will be greater if his animation is  'PARACHUTE    FALL_SKYDIVE`. The above made changes in the include overcomes this bug. Thus, it needed the `CRF\c_stclib` to be improved a little bit.