Add WC_YSI_PACKAGE (true/false) macro function
This is only for people who not using the YSI  packages, like me so i made this easy, the  scripters before including the Weapon Config can set with code:
#define WC_YSI_PACKAGE false, and disable Foreach iterate function and '#error the YSI library is required, get it here: github.com/pawn-lang/YSI-Includes' error.
The YSI Libary is by default set to enabled (true).
Y-Less i respect your work this is not personal, but someone don't want to include all package via 1 include. Thats it, thanks.